Add files via upload
The Library Management System is a console-based application developed in C# as part of the Web Engineering course (CSF21M). This project aims to facilitate the management of library resources and enhance the efficiency of library operations. The system provides functionalities for managing books, patrons, and transactions within a library environment.